Move macro implementation to a separate crate
