+++ crates/node-bindings/Cargo.toml
@@ -17,6 +17,7 @@ xxhash-rust = { version = "0.8.2", features = ["xxh3"] }
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 napi = {version =  "2.12.6", features = ["serde-json", "napi4", "napi5"]}
 parcel-dev-dep-resolver = { path = "../../packages/utils/dev-dep-resolver" }
+parcel-macros = { path = "../macros", features = ["napi"] }
 oxipng = "8.0.0"
 mozjpeg-sys = "1.0.0"
 libc = "0.2"
