+++ packages/transformers/js/core/Cargo.toml
@@ -34,3 +34,4 @@ dunce = "1.0.1"
 pathdiff = "0.2.0"
 path-slash = "0.1.4"
 indexmap = "1.9.2"
+parcel-macros = { path = "../../../../crates/macros" }
